requirements.txt
Flask
Pillow
wave
numpy
opencv-python
stegano
pydub